Refactor map and geo helpers into packages
