+++ src/fedviz/integrations/__init__.py
@@ -1,0 +1,3 @@
+from .grpc_geo import patch_communicator_for_geo
+
+__all__ = ["patch_communicator_for_geo"]
