+++ src/fedviz/map/__init__.py
@@ -1,0 +1,9 @@
+from .metadata import build_map_metadata_from_events, build_rounds_from_events, merge_client_state
+from .server import MapServer
+
+__all__ = [
+    "MapServer",
+    "build_map_metadata_from_events",
+    "build_rounds_from_events",
+    "merge_client_state",
+]
